changing int to float

+++ module-1/addition_subtraction.py
@@ -1,6 +1,6 @@
 print('Enter first number: ', end = ' ')
-num1 = int(input())
+num1 = float(input())
 print('Enter second number: ', end = ' ')
-num2 = int(input())
+num2 = float(input())
 print('Sum of the 2 numbers: ', num1 + num2)
 print('Difference of the 2 numbers: ', num1 - num2)
